+++ SourceCode/GPS/Classes/CTrack.cs
@@ -318,7 +318,7 @@ namespace AgOpenGPS
         public vec2 endPtB;
         public TrackMode mode;
         public double nudgeDistance;
-        public List<double> workedTracks = new List<double>();
+        public List<int> workedTracks = new List<int>();
 
         public CTrk()
         {
